Converge GameService tests into a single test file
